fix: Fix DML ON clause for imperative statements
